Bug fix: accounts list is not working properly.
The accounts list is read from local storage `boot.accounts` array. Prior to this commit this array is only filled during authorisation (login), where next-gen infers the list of accounts from the response url. This array needs to be also updated when accounts are created (virtual, real, malta); ergo the intro of this `addAccountToBoot` method.

Updating the local storage directly instead of a method that updates both the state and the local storage is harmless in this context, since upon adding a new account the page is always refreshed, thereby copying the values in the local storage to the state.