bug fix: hamburger doesn't respond when clicked.
Seems to only happen in development (localhost and running cordova in debug mode), which can be annoying. I did a quick read up on why on the [original fastclick project](https://github.com/ftlabs/fastclick); turns out there a 300ms delay from the time you tap the button to fire a click event, but, for newer browsers this is no issue:

> Note: As of late 2015 most mobile browsers - notably Chrome and Safari - no longer have a 300ms touch delay, so fastclick offers no benefit on newer browsers, and risks introducing bugs into your application. Consider carefully whether you really need to use it.

So, no point having this extension if it introduces this bug.